Autoplot method for bench_mark objects

Autoplot method for bench_mark objects

autoplot.bench_mark(object, type = c("beeswarm", "jitter", "ridge", "boxplot",
  "violin"), ...)

# S3 method for bench_mark
plot(x, ..., type = c("beeswarm", "jitter", "ridge",
  "boxplot", "violin"), y)

Arguments

object
A bench_mark object.

type
The type of plot. Plotting geoms used for each type are

beeswarm - ggbeeswarm::geom_quasirandom()

jitter - ggplot2::geom_jitter()

ridge - ggridges::geom_density_ridges()

boxplot - ggplot2::geom_boxplot()

violin - ggplot2::geom_violin()

...
Additional arguments passed to the plotting geom.

x
A bench_mark object.

y
Ignored, required for compatibility with the plot() generic.

Details

This function requires some optional dependencies. ggplot2, tidyr, and depending on the plot type ggbeeswarm, ggridges.

For type of beeswarm and jitter the points are colored by the highest level garbage collection performed during each iteration.

For plots with 2 parameters ggplot2::facet_grid() is used to construct a 2d facet. For other numbers of parameters ggplot2::facet_wrap() is used instead.