bench: High Precision Timing of R Expressions

Tools to accurately benchmark and analyze execution times for R expressions.

See also

Examples

dat <- data.frame(x = runif(10000, 1, 1000), y=runif(10000, 1, 1000))

# `bench::mark()` implicitly calls summary() automatically
results <- bench::mark(
  dat[dat$x > 500, ],
  dat[which(dat$x > 500), ],
  subset(dat, x > 500))

# However you can also do so explicitly to filter gc differently.
summary(results, filter_gc = FALSE)

#> # A tibble: 3 x 13
#>   expression                  min median `itr/sec` mem_alloc `gc/sec` n_itr
#>   <bch:expr>                <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int>
#> 1 dat[dat$x > 500, ]        399µs  421µs     1569.     374KB     12.0   785
#> 2 dat[which(dat$x > 500), ] 306µs  321µs     2794.     258KB     16.0  1397
#> 3 subset(dat, x > 500)      523µs  545µs     1655.     492KB     16.0   828
#> # … with 6 more variables: n_gc <dbl>, total_time <bch:tm>, result <list>,
#> #   memory <list>, time <list>, gc <list>

# Or output relative times
summary(results, relative = TRUE)

#> # A tibble: 3 x 13
#>   expression                  min median `itr/sec` mem_alloc `gc/sec` n_itr
#>   <bch:expr>                <dbl>  <dbl>     <dbl>     <dbl>    <dbl> <int>
#> 1 dat[dat$x > 500, ]         1.30   1.31      1.25      1.45     1      779
#> 2 dat[which(dat$x > 500), ]  1      1         1.69      1        1.01  1389
#> 3 subset(dat, x > 500)       1.71   1.70      1         1.91     1.01   820
#> # … with 6 more variables: n_gc <dbl>, total_time <bch:tm>, result <list>,
#> #   memory <list>, time <list>, gc <list>