Measure Process CPU and real time that an expression used.

Measure Process CPU and real time that an expression used.

Usage

bench_time(expr)

Arguments

expr

A expression to be timed.

Value

A bench_time object with two values.

process - The process CPU usage of the expression evaluation.

real - The wallclock time of the expression evaluation.

Details

On some systems (such as macOS) the process clock has lower precision than the realtime clock, as a result there may be cases where the process time is larger than the real time for fast expressions.

See also

Examples

# This will use ~.5 seconds of real time, but very little process time.
bench_time(Sys.sleep(.5))
#> process    real 
#>  64.5µs 500.6ms