Custom printing function for bench_mark objects in knitr documents

By default, data columns (result, memory, time, gc) are omitted when printing in knitr. If you would like to include these columns, set the knitr chunk option bench.all_columns = TRUE.

Usage

knit_print.bench_mark(x, ..., options)

Arguments

x

An R object to be printed

...

Additional arguments passed to the S3 method. Currently ignored, except two optional arguments options and inline; see the references below.

options

A list of knitr chunk options set in the currently evaluated chunk.

Details

You can set bench.all_columns = TRUE to show all columns of the bench mark object.

```{r, bench.all_columns = TRUE}
bench::mark(
  subset(mtcars, cyl == 3),
  mtcars[mtcars$cyl == 3, ]
)
```