Summarize bench::mark results.

Summarize bench::mark results.

# S3 method for bench_mark
summary(object, filter_gc = TRUE, relative = FALSE,
  ...)

Arguments

object
bench_mark object to summarize.

filter_gc
If TRUE filter iterations that contained at least one garbage collection before summarizing.

relative
If TRUE all summaries are computed relative to the minimum execution time rather than absolute time.

...
Additional arguments ignored.

Value

A tibble with the additional summary columns. The following summary columns are computed

min - bench_time The minimum execution time.

mean - bench_time The arithmetic mean of execution time

median - bench_time The sample median of execution time.

max - bench_time The maximum execution time.

mem_alloc - bench_bytes Total amount of memory allocated by running the expression.

itr/sec - integer The estimated number of executions performed per second.

n_itr - integer Total number of iterations after filtering garbage collections (if filter_gc == TRUE).

n_gc - integer Total number of garbage collections performed over all runs.

Details

If filter_gc == TRUE (the default) runs that contain a garbage collection will be removed before summarizing. This is most useful for fast expressions when the majority of runs do not contain a gc. Call summary(filter_gc = FALSE) if you would like to compute summaries with these times, such as expressions with lots of allocations when all or most runs contain a gc.

Examples

dat <- data.frame(x = runif(10000, 1, 1000), y=runif(10000, 1, 1000))

# `bench::mark()` implicitly calls summary() automatically
results <- bench::mark(
  dat[dat$x > 500, ],
  dat[which(dat$x > 500), ],
  subset(dat, x > 500))

# However you can also do so explicitly to filter gc differently.
summary(results, filter_gc = FALSE)

#> # A tibble: 3 x 14
#>   expression   min  mean median    max `itr/sec` mem_alloc  n_gc n_itr
#>   <chr>      <bch> <bch> <bch:> <bch:>     <dbl> <bch:byt> <dbl> <int>
#> 1 dat[dat$x… 433µs 509µs  443µs 5.45ms     1963.     375KB    12   981
#> 2 dat[which… 378µs 438µs  385µs 5.12ms     2281.     258KB     9  1139
#> 3 subset(da… 546µs 624µs  557µs 5.17ms     1603.     493KB    10   801
#> # ... with 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#> #   time <list>, gc <list>

# Or output relative times
summary(results, relative = TRUE)

#> # A tibble: 3 x 14
#>   expression   min  mean median   max `itr/sec` mem_alloc  n_gc n_itr total_time
#>   <chr>      <dbl> <dbl>  <dbl> <dbl>     <dbl>     <dbl> <dbl> <dbl>      <dbl>
#> 1 dat[dat$x…  1.15  1.13   1.15  1         1.24      1.45  1.33  1.23       1   
#> 2 dat[which…  1     1      1     3.44      1.41      1     1     1.43       1.03
#> 3 subset(da…  1.45  1.41   1.45  1.15      1         1.91  1.11  1          1.01
#> # ... with 4 more variables: result <list>, memory <list>, time <list>,
#> #   gc <list>