bench: High Precision Timing of R Expressions

Tools to accurately benchmark and analyze execution times for R expressions.

See also

Author

Maintainer: Davis Vaughan davis@posit.co

Authors:

Jim Hester

Other contributors:

Drew Schmidt (read_proc_file implementation) [contributor]

Posit Software, PBC [copyright holder, funder]

Examples

dat <- data.frame(x = runif(10000, 1, 1000), y=runif(10000, 1, 1000))

# `bench::mark()` implicitly calls summary() automatically
results <- bench::mark(
  dat[dat$x > 500, ],
  dat[which(dat$x > 500), ],
  subset(dat, x > 500))

# However you can also do so explicitly to filter gc differently.
summary(results, filter_gc = FALSE)
#> # A tibble: 3 × 13
#>   expression           min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr>         <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>
#> 1 dat[dat$x > 500, ] 148µs  156µs     3652.     374KB     32.0  1826    16
#> 2 dat[which(dat$x >… 133µs  141µs     4708.     258KB     28.0  2354    14
#> 3 subset(dat, x > 5… 186µs  212µs     2794.     492KB     34.0  1397    17
#> # ℹ 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#> #   time <list>, gc <list>

# Or output relative times
summary(results, relative = TRUE)
#> # A tibble: 3 × 13
#>   expression           min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#>   <bch:expr>         <dbl>  <dbl>     <dbl>     <dbl>    <dbl> <int> <dbl>
#> 1 dat[dat$x > 500, ]  1.11   1.11      1.37      1.45     1.33  1810    16
#> 2 dat[which(dat$x >…  1      1         1.51      1        1     2340    14
#> 3 subset(dat, x > 5…  1.40   1.50      1         1.91     1.36  1380    17
#> # ℹ 5 more variables: total_time <bch:tm>, result <list>, memory <list>,
#> #   time <list>, gc <list>