bench: High Precision Timing of R Expressions

Tools to accurately benchmark and analyze execution times for R expressions.

See also

Examples

dat <- data.frame(x = runif(10000, 1, 1000), y=runif(10000, 1, 1000))

# `bench::mark()` implicitly calls summary() automatically
results <- bench::mark(
  dat[dat$x > 500, ],
  dat[which(dat$x > 500), ],
  subset(dat, x > 500))

# However you can also do so explicitly to filter gc differently.
summary(results, filter_gc = FALSE)

#> # A tibble: 3 x 13
#>   expression                  min median `itr/sec` mem_alloc `gc/sec` n_itr
#>   <bch:expr>                <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int>
#> 1 dat[dat$x > 500, ]        332µs  353µs     2097.     374KB     16.0  1051
#> 2 dat[which(dat$x > 500), ] 253µs  266µs     3382.     258KB     16.0  1691
#> 3 subset(dat, x > 500)      441µs  465µs     1955.     492KB     20.0   978
#> # … with 6 more variables: n_gc <dbl>, total_time <bch:tm>, result <list>,
#> #   memory <list>, time <list>, gc <list>

# Or output relative times
summary(results, relative = TRUE)

#> # A tibble: 3 x 13
#>   expression                  min median `itr/sec` mem_alloc `gc/sec` n_itr
#>   <bch:expr>                <dbl>  <dbl>     <dbl>     <dbl>    <dbl> <int>
#> 1 dat[dat$x > 500, ]         1.31   1.32      1.29      1.45     1.22  1043
#> 2 dat[which(dat$x > 500), ]  1      1         1.72      1        1     1683
#> 3 subset(dat, x > 500)       1.74   1.74      1         1.91     1.27   968
#> # … with 6 more variables: n_gc <dbl>, total_time <bch:tm>, result <list>,
#> #   memory <list>, time <list>, gc <list>