bench: High Precision Timing of R Expressions

Tools to accurately benchmark and analyze execution times for R expressions.

See also

Author

Maintainer: Davis Vaughan davis@posit.co

Authors:

Jim Hester

Other contributors:

Drew Schmidt (read_proc_file implementation) [contributor]

Posit Software, PBC [copyright holder, funder]

Examples

dat <- data.frame(x = runif(10000, 1, 1000), y=runif(10000, 1, 1000))

# `bench::mark()` implicitly calls summary() automatically
results <- bench::mark(
  dat[dat$x > 500, ],
  dat[which(dat$x > 500), ],
  subset(dat, x > 500))

# However you can also do so explicitly to filter gc differently.
summary(results, filter_gc = FALSE)
#> # A tibble: 3 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>    <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 dat[dat$x > … 298µs  306µs     2381.     376KB     17.2  1246     9      523ms
#> 2 dat[which(da… 224µs  231µs     3450.     259KB     16.0  1725     8      500ms
#> 3 subset(dat, … 413µs  431µs     1850.     493KB     16.0   925     8      500ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>

# Or output relative times
summary(results, relative = TRUE)
#> # A tibble: 3 × 13
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr>    <dbl>  <dbl>     <dbl>     <dbl>    <dbl> <int> <dbl>   <bch:tm>
#> 1 dat[dat$x > …  1.33   1.32      1.36      1.45     1.14  1237     9      396ms
#> 2 dat[which(da…  1      1         1.86      1        1     1717     8      402ms
#> 3 subset(dat, …  1.85   1.86      1         1.91     1.01   917     8      399ms
#> # ℹ 4 more variables: result <list>, memory <list>, time <list>, gc <list>