Benchmark a series of functions

Benchmark a list of quoted expressions. Each expression will always run at least twice, once to measure the memory allocation and store results and one or more times to measure timing.

mark(..., setup = NULL, parameters = list(), env = parent.frame(),
  min_time = 0.5, iterations = NULL, min_iterations = 1,
  max_iterations = 10000, check = TRUE)

Arguments

...
Expressions to benchmark

setup
code to evaluated before each benchmark group, this code will be reevaluated for each row in parameters.

parameters
Variable values to assign, all values will be enumerated by expand.grid().

env
The environment which to evaluate the expressions

min_time
The minimum number of seconds to run each expression, set to Inf to always run max_iterations times instead.

iterations
If not NULL, the default, run each expression for exactly this number of iterations. This overrides both min_iterations and max_iterations.

min_iterations
Each expression will be evaluated a minimum of min_iterations times.

max_iterations
Each expression will be evaluated a maximum of max_iterations times.

check
Check if results are consistent. If TRUE, checking is done with all.equal(), if FALSE checking is disabled. If check is a function that function will be called with each pair of results to determine consistency.

Value

A tibble with the additional summary columns. The following summary columns are computed

min - bench_time The minimum execution time.

mean - bench_time The arithmetic mean of execution time

median - bench_time The sample median of execution time.

max - bench_time The maximum execution time.

mem_alloc - bench_bytes Total amount of memory allocated by running the expression.

itr/sec - integer The estimated number of executions performed per second.

n_itr - integer Total number of iterations after filtering garbage collections (if filter_gc == TRUE).

n_gc - integer Total number of garbage collections performed over all runs.

Examples

mark(
  setup = dat <- data.frame(x = runif(num_x, 1, 1000), y=runif(num_y, 1, 1000)),
  parameters = list(num_x = c(1000, 10000), num_y = c(1000, 10000)),
  min_time = .1,

  dat[dat$x > 500, ],
  dat[which(dat$x > 500), ],
  subset(dat, x > 500))

#> Running benchmark with:
#>   num_x num_y

#> 1  1000  1000

#> 2 10000  1000

#> 3  1000 10000

#> 4 10000 10000

#> # A tibble: 12 x 16
#>    expression  num_x num_y     min    mean  median      max `itr/sec` mem_alloc
#>    <chr>       <dbl> <dbl> <bch:t> <bch:t> <bch:t> <bch:tm>     <dbl> <bch:byt>
#>  1 dat[dat$x …  1000  1000  46.2µs  72.8µs  56.3µs  941.5µs    13738.     39.5K
#>  2 dat[which(…  1000  1000  45.8µs  56.9µs  51.9µs   1.01ms    17561.     33.5K
#>  3 subset(dat…  1000  1000  64.7µs  88.9µs  74.5µs   1.07ms    11244.     51.3K
#>  4 dat[dat$x … 10000  1000 318.6µs 427.8µs 378.5µs    1.3ms     2338.    414.2K
#>  5 dat[which(… 10000  1000 252.9µs 323.4µs 290.5µs   1.05ms     3092.    355.4K
#>  6 subset(dat… 10000  1000 399.8µs 486.4µs 460.8µs   1.35ms     2056.    531.5K
#>  7 dat[dat$x …  1000 10000 221.8µs 299.9µs 255.8µs   1.26ms     3335.    415.7K
#>  8 dat[which(…  1000 10000 214.1µs 283.4µs 250.4µs   1.09ms     3528.    357.1K
#>  9 subset(dat…  1000 10000 259.7µs 321.7µs   299µs 839.84µs     3108.      533K
#> 10 dat[dat$x … 10000 10000 323.4µs 416.7µs 368.9µs   1.23ms     2400.    416.6K
#> 11 dat[which(… 10000 10000 251.6µs 320.8µs 288.3µs   1.18ms     3118.      358K
#> 12 subset(dat… 10000 10000 406.3µs 489.5µs   459µs   1.36ms     2043.    533.9K
#> # ... with 7 more variables: n_gc <dbl>, n_itr <int>, total_time <bch:tm>,
#> #   result <list>, memory <list>, time <list>, gc <list>