Run setup code and benchmarks across a grid of parameters

press() is used to run bench::mark() across a grid of parameters and then press the results together.

The parameters you want to set are given as named arguments and a grid of all possible combinations is automatically created.

The code to setup and benchmark is given by one unnamed expression (often delimited by \{).

If replicates are desired a dummy variable can be used, e.g. rep = 1:5 for replicates.

press(..., .grid = NULL)

Arguments

...
If named, parameters to define, if unnamed the expression to run. Only one unnamed expression is permitted.

.grid
A pre-build grid of values to use, typically a data.frame or tibble. This is useful if you only want to use a subset of all possible combinations.

Examples

# Helper function to create a simple data.frame of the specified dimensions
create_df <- function(rows, cols) {
  as.data.frame(setNames(
    replicate(cols, runif(rows, 1, 1000), simplify = FALSE),
    rep_len(c("x", letters), cols)))
}

# Run 4 data sizes across 3 samples with 2 replicates (24 total benchmarks)
press(
  rows = c(1000, 10000),
  cols = c(10, 100),
  rep = 1:2,
  {
    dat <- create_df(rows, cols)
    bench::mark(
      min_time = .05,
      bracket = dat[dat$x > 500, ],
      which = dat[which(dat$x > 500), ],
      subset = subset(dat, x > 500)
    )
  }
)

#> Running with:
#>    rows  cols   rep

#> 1  1000    10     1

#> 2 10000    10     1

#> 3  1000   100     1

#> 4 10000   100     1

#> 5  1000    10     2

#> 6 10000    10     2

#> 7  1000   100     2

#> 8 10000   100     2

#> # A tibble: 24 x 16
#>    expression  rows  cols   rep      min   median `itr/sec` mem_alloc `gc/sec`
#>    <bch:expr> <dbl> <dbl> <int> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#>  1 bracket     1000    10     1  92.06µs 139.26µs     7565.  112.77KB      0  
#>  2 which       1000    10     1  81.18µs  88.09µs    10296.   53.87KB     22.7
#>  3 subset      1000    10     1 117.44µs 127.75µs     7434.  124.63KB     22.7
#>  4 bracket    10000    10     1 822.69µs 849.97µs     1160.    1.14MB     22.3
#>  5 which      10000    10     1  404.4µs 425.86µs     2125.  575.37KB     21.9
#>  6 subset     10000    10     1 943.98µs 971.66µs     1015.    1.25MB     49.5
#>  7 bracket     1000   100     1 735.73µs 789.89µs     1258.  987.79KB     22.9
#>  8 which       1000   100     1 610.27µs 665.28µs     1479.  398.73KB     22.7
#>  9 subset      1000   100     1 805.88µs 851.65µs     1161. 1008.51KB     22.8
#> 10 bracket    10000   100     1   6.77ms   6.79ms      147.    9.72MB    295. 
#> # … with 14 more rows, and 7 more variables: n_itr <int>, n_gc <dbl>,
#> #   total_time <bch:tm>, result <list>, memory <list>, time <list>, gc <list>