Run setup code and benchmarks across a grid of parameters

press() is used to run bench::mark() across a grid of parameters and then press the results together.

The parameters you want to set are given as named arguments and a grid of all possible combinations is automatically created.

The code to setup and benchmark is given by one unnamed expression (often delimited by {).

If replicates are desired a dummy variable can be used, e.g. rep = 1:5 for replicates.

press(..., .grid = NULL)

Arguments

...
If named, parameters to define, if unnamed the expression to run. Only one unnamed expression is permitted.

.grid
A pre-build grid of values to use, typically a data.frame or tibble. This is useful if you only want to use a subset of all possible combinations.

Examples

# Helper function to create a simple data.frame of the specified dimensions
create_df <- function(rows, cols) {
  as.data.frame(setNames(
    replicate(cols, runif(rows, 1, 1000), simplify = FALSE),
    rep_len(c("x", letters), cols)))
}

# Run 4 data sizes across 3 samples with 2 replicates (24 total benchmarks)
press(
  rows = c(1000, 10000),
  cols = c(10, 100),
  rep = 1:2,
  {
    dat <- create_df(rows, cols)
    bench::mark(
      min_time = .05,
      bracket = dat[dat$x > 500, ],
      which = dat[which(dat$x > 500), ],
      subset = subset(dat, x > 500)
    )
  }
)

#> Running with:
#>    rows  cols   rep

#> 1  1000    10     1

#> 2 10000    10     1

#> 3  1000   100     1

#> 4 10000   100     1

#> 5  1000    10     2

#> 6 10000    10     2

#> 7  1000   100     2

#> 8 10000   100     2

#> # A tibble: 24 x 16
#>    expression  rows  cols   rep      min   median `itr/sec` mem_alloc `gc/sec`
#>    <bch:expr> <dbl> <dbl> <int> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#>  1 bracket     1000    10     1 147.72µs 167.33µs     5504.  112.77KB     21.1
#>  2 which       1000    10     1  96.63µs 103.24µs     9371.   53.87KB      0  
#>  3 subset      1000    10     1 175.56µs  243.5µs     4357.  124.63KB     23.8
#>  4 bracket    10000    10     1  961.4µs 979.25µs     1011.    1.14MB     23.0
#>  5 which      10000    10     1 479.69µs 496.44µs     1894.  575.37KB     22.6
#>  6 subset     10000    10     1    1.1ms   1.11ms      890.    1.25MB     22.2
#>  7 bracket     1000   100     1   1.17ms   1.21ms      819.  987.79KB     23.4
#>  8 which       1000   100     1 730.86µs 764.07µs     1298.  398.73KB     23.6
#>  9 subset      1000   100     1   1.24ms    1.3ms      768. 1008.51KB     23.3
#> 10 bracket    10000   100     1   7.73ms   7.87ms      127.    9.72MB    191. 
#> # … with 14 more rows, and 7 more variables: n_itr <int>, n_gc <dbl>,
#> #   total_time <bch:tm>, result <list>, memory <list>, time <list>, gc <list>