Run setup code and benchmarks across a grid of parameters

press() is used to run mark() across a grid of parameters and then press the results together.

The parameters you want to set are given as named arguments and a grid of all possible combinations is automatically created.

The code to setup and benchmark is given by one unnamed expression (often delimited by \{).

If replicates are desired a dummy variable can be used, e.g. rep = 1:5 for replicates.

press(..., .grid = NULL)

Arguments

...

If named, parameters to define, if unnamed the expression to run. Only one unnamed expression is permitted.

.grid

A pre-built grid of values to use, typically a data.frame or tibble. This is useful if you only want to benchmark a subset of all possible combinations.

Examples

# Helper function to create a simple data.frame of the specified dimensions
create_df <- function(rows, cols) {
  as.data.frame(setNames(
    replicate(cols, runif(rows, 1, 1000), simplify = FALSE),
    rep_len(c("x", letters), cols)))
}

# Run 4 data sizes across 3 samples with 2 replicates (24 total benchmarks)
press(
  rows = c(1000, 10000),
  cols = c(10, 100),
  rep = 1:2,
  {
    dat <- create_df(rows, cols)
    bench::mark(
      min_time = .05,
      bracket = dat[dat$x > 500, ],
      which = dat[which(dat$x > 500), ],
      subset = subset(dat, x > 500)
    )
  }
)
#> Running with:
#>    rows  cols   rep
#> 1  1000    10     1
#> 2 10000    10     1
#> 3  1000   100     1
#> 4 10000   100     1
#> 5  1000    10     2
#> 6 10000    10     2
#> 7  1000   100     2
#> 8 10000   100     2
#> # A tibble: 24 × 16
#>    expression  rows  cols   rep      min   median `itr/sec` mem_alloc `gc/sec`
#>    <bch:expr> <dbl> <dbl> <int> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#>  1 bracket     1000    10     1  118.7µs  124.3µs     7855.  113.27KB      0  
#>  2 which       1000    10     1  110.4µs  145.6µs     7015.   54.21KB     30.6
#>  3 subset      1000    10     1  154.7µs  161.9µs     6053.  125.13KB      0  
#>  4 bracket    10000    10     1 953.81µs 974.01µs     1014.    1.13MB     32.7
#>  5 which      10000    10     1  455.4µs  465.5µs     1990.  570.98KB     28.8
#>  6 subset     10000    10     1    1.1ms   1.12ms      880.    1.25MB     29.3
#>  7 bracket     1000   100     1      1ms   1.02ms      975. 1009.77KB     27.9
#>  8 which       1000   100     1 830.81µs 865.51µs     1150.   412.9KB      0  
#>  9 subset      1000   100     1   1.06ms   1.09ms      903.    1.01MB     30.1
#> 10 bracket    10000   100     1   7.43ms   7.61ms      132.    9.72MB     88.1
#> # ℹ 14 more rows
#> # ℹ 7 more variables: n_itr <int>, n_gc <dbl>, total_time <bch:tm>,
#> #   result <list>, memory <list>, time <list>, gc <list>