Run setup code and benchmarks across a grid of parameters

press() is used to run mark() across a grid of parameters and then press the results together.

The parameters you want to set are given as named arguments and a grid of all possible combinations is automatically created.

The code to setup and benchmark is given by one unnamed expression (often delimited by \{).

If replicates are desired a dummy variable can be used, e.g. rep = 1:5 for replicates.

Usage

press(..., .grid = NULL, .quiet = FALSE)

Arguments

...

If named, parameters to define, if unnamed the expression to run. Only one unnamed expression is permitted.

.grid

A pre-built grid of values to use, typically a data.frame() or tibble::tibble(). This is useful if you only want to benchmark a subset of all possible combinations.

.quiet

If TRUE, progress messages will not be emitted.

Examples

# Helper function to create a simple data.frame of the specified dimensions
create_df <- function(rows, cols) {
  as.data.frame(setNames(
    replicate(cols, runif(rows, 1, 1000), simplify = FALSE),
    rep_len(c("x", letters), cols)))
}

# Run 4 data sizes across 3 samples with 2 replicates (24 total benchmarks)
press(
  rows = c(1000, 10000),
  cols = c(10, 100),
  rep = 1:2,
  {
    dat <- create_df(rows, cols)
    bench::mark(
      min_time = .05,
      bracket = dat[dat$x > 500, ],
      which = dat[which(dat$x > 500), ],
      subset = subset(dat, x > 500)
    )
  }
)
#> Running with:
#>    rows  cols   rep
#> 1  1000    10     1
#> 2 10000    10     1
#> 3  1000   100     1
#> 4 10000   100     1
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> 5  1000    10     2
#> 6 10000    10     2
#> 7  1000   100     2
#> 8 10000   100     2
#> # A tibble: 24 × 16
#>    expression  rows  cols   rep      min   median `itr/sec` mem_alloc
#>    <bch:expr> <dbl> <dbl> <int> <bch:tm> <bch:tm>     <dbl> <bch:byt>
#>  1 bracket     1000    10     1   98.1µs  102.2µs   9405.    112.77KB
#>  2 which       1000    10     1   89.5µs   93.7µs  10474.     53.87KB
#>  3 subset      1000    10     1  121.7µs  127.4µs   7578.    124.63KB
#>  4 bracket    10000    10     1  329.9µs  346.2µs   2760.      1.14MB
#>  5 which      10000    10     1  262.7µs    278µs   3451.    575.37KB
#>  6 subset     10000    10     1  389.3µs  412.3µs   2306.      1.25MB
#>  7 bracket     1000   100     1  772.7µs  804.7µs   1222.    986.96KB
#>  8 which       1000   100     1  685.4µs  705.8µs   1410.     397.9KB
#>  9 subset      1000   100     1  823.3µs    841µs   1152.   1008.18KB
#> 10 bracket    10000   100     1  134.2ms  134.2ms      7.45    9.71MB
#> # ℹ 14 more rows
#> # ℹ 8 more variables: `gc/sec` <dbl>, n_itr <int>, n_gc <dbl>,
#> #   total_time <bch:tm>, result <list>, memory <list>, time <list>,
#> #   gc <list>