Workout a group of expressions individually

Given an block of expressions in {} workout() individually times each expression in the group. workout_expressions() is a lower level function most useful when reading lists of calls from a file.

workout(expr, description = NULL)

workout_expressions(exprs, env = parent.frame(), description = NULL)

Arguments

expr

one or more expressions to workout, use {} to pass multiple expressions.

description

A name to label each expression, if not supplied the deparsed expression will be used.

exprs

A list of calls to measure.

env

The environment in which the expressions should be evaluated.

Examples

workout({
  x <- 1:1000
  evens <- x %% 2 == 0
  y <- x[evens]
  length(y)
  length(which(evens))
  sum(evens)
})
#> # A tibble: 6 × 3
#>   exprs                 process     real
#>   <bch:expr>           <bch:tm> <bch:tm>
#> 1 x <- 1:1000             5.4µs   9.06µs
#> 2 evens <- x%%2 == 0     38.7µs  40.29µs
#> 3 y <- x[evens]          10.1µs  11.21µs
#> 4 length(y)               1.5µs   2.62µs
#> 5 length(which(evens))    9.7µs  10.73µs
#> 6 sum(evens)              3.5µs   4.53µs

# The equivalent to the above, reading the code from a file
workout_expressions(as.list(parse(system.file("examples/exprs.R", package = "bench"))))
#> # A tibble: 6 × 3
#>   exprs                 process     real
#>   <bch:expr>           <bch:tm> <bch:tm>
#> 1 x <- 1:1000             3.5µs   5.72µs
#> 2 evens <- x%%2 == 0     31.8µs   32.9µs
#> 3 y <- x[evens]             9µs  10.25µs
#> 4 length(y)               1.5µs   2.62µs
#> 5 length(which(evens))    8.3µs    9.3µs
#> 6 sum(evens)              2.9µs   4.29µs