Workout a group of expressions individually

Given an block of expressions in {} workout() individually times each expression in the group. workout_expressions() is a lower level function most useful when reading lists of calls from a file.

workout(expr, description = NULL)

workout_expressions(exprs, env = parent.frame(), description = NULL)

Arguments

expr
one or more expressions to workout, use {} to pass multiple expressions.

description
A name to label each expression, if not supplied the deparsed expression will be used.

exprs
A list of calls to measure.

env
The environment in which the expressions should be evaluated.

Examples

workout({
  x <- 1:1000
  evens <- x %% 2 == 0
  y <- x[evens]
  length(y)
  length(which(evens))
  sum(evens)
})

#> # A tibble: 6 x 3
#>   exprs                 process     real
#>   <bch:expr>           <bch:tm> <bch:tm>
#> 1 x <- 1:1000            3.25µs   5.96µs
#> 2 evens <- x%%2 == 0    41.07µs  41.48µs
#> 3 y <- x[evens]          5.63µs   5.96µs
#> 4 length(y)              1.02µs   1.43µs
#> 5 length(which(evens))   4.57µs   5.01µs
#> 6 sum(evens)             2.66µs   2.86µs

# The equivalent to the above, reading the code from a file
workout_expressions(as.list(parse(system.file("examples/exprs.R", package = "bench"))))

#> # A tibble: 6 x 3
#>   exprs                 process     real
#>   <bch:expr>           <bch:tm> <bch:tm>
#> 1 x <- 1:1000            2.99µs   4.77µs
#> 2 evens <- x%%2 == 0     39.1µs  39.58µs
#> 3 y <- x[evens]          5.05µs   5.25µs
#> 4 length(y)               615ns 953.67ns
#> 5 length(which(evens))   4.87µs   5.48µs
#> 6 sum(evens)             2.53µs   2.86µs